[accessibility] update TogetherNet capitalization
Access report details provided by Walei Sabry:

Spelling of Togethernet
When it is spelled Togethernet, screen readers do not pronounce it correctly. Suggestion to always write it as “TogetherNet.” Not only will it be read by screen readers correctly, but it will make sense to more folks. Note: do not forget to update the name in alt text as well.