added anchor links to footnotes
Accessibility repot by Walei Sabry:

Footnotes and back to top links
Status: not resolved
3/20 update: No changes

Initial Issue:
The footnote links are accessible and take screen reader focus to the appropriate area. However, once the user is in that area, it will be difficult to navigate back to the spot they were reading on the page. Suggestion to add anchor links in the End notes section that take the user back to each of the citations in the text. For example, if a user activates the first end note, they will be taken to the end notes area at the bottom of the  page. After reading the first end note, it would be great if there was a link that can take the user back to the spot where end note 1 is in the text. That way users  can more conveniently navigate back and forth while reading end notes.