detailed alt texts
Accessibility report by Walei Sabry:

More Detailed Alt-text
Status: not resolved
3/20 update: alt-text has not changed.

Initial issue:
Suggestion to make the alt-text a little more  detailed. For example, include a few details about the TogetherNet logo. What colors is it? For screen shots, it would be good to include some of the details on screen or a brief description of the layout of that screen.